+++ tests/testthat/test-getEvents.R
@@ -4,7 +4,7 @@ test_that("makes expected HTTP request", {
   with_mock(
     `httr::request_perform` = function(req, handle, refresh) {
       expect_equal(req$method, "GET")
-      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.FCS?compensatedQ=FALSE&headers=TRUE&addEventNumber=TRUE") # nolint
+      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.FCS?populationId=null&compensationId=0&compensatedQ=false&headers=true&addEventNumber=true") # nolint
       response <- httptest::fake_response(
         req$url,
         req$method,
@@ -25,7 +25,7 @@ test_that("makes expected HTTP request with subsampling", {
   with_mock(
     `httr::request_perform` = function(req, handle, refresh) {
       expect_equal(req$method, "GET")
-      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.FCS?compensatedQ=FALSE&headers=TRUE&addEventNumber=FALSE&preSubsampleN=50&seed=2.25") # nolint
+      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.FCS?populationId=null&compensationId=0&compensatedQ=false&headers=true&addEventNumber=false&preSubsampleN=50&seed=2.25") # nolint
       response <- httptest::fake_response(
         req$url,
         req$method,
@@ -50,7 +50,7 @@ test_that("doesn't mangle column names", {
   with_mock(
     `httr::request_perform` = function(req, handle, refresh) {
       expect_equal(req$method, "GET")
-      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.TSV?compensatedQ=FALSE&headers=TRUE&addEventNumber=FALSE") # nolint
+      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.TSV?populationId=null&compensationId=0&compensatedQ=false&headers=true&addEventNumber=false") # nolint
       response <- httptest::fake_response(
         req$url,
         req$method,
@@ -64,6 +64,54 @@ test_that("doesn't mangle column names", {
       setServer("https://my.server.com")
       resp <- getEvents("591a3b441d725115208a6fda", "591a3b441d725115208a6fdc", format = "TSV")
       expect_equal(colnames(resp), c("FSC-A", "CD3 (Ax647-A)"))
+    }
+  )
+})
+
+test_that("makes expected HTTP request for S3 transfer", {
+  with_mock(
+    `httr::request_perform` = function(req, handle, refresh) {
+      expect_equal(req$method, "POST")
+      expect_equal(req$url, "https://my.server.com/api/v1/experiments/591a3b441d725115208a6fda/fcsfiles/591a3b441d725115208a6fdc.FCS?populationId=null&compensationId=0&compensatedQ=false&headers=true&addEventNumber=false") # nolint
+      body <- rawToChar(req$options$postfields)
+      expect_equal(body, '{"dest":{"host":"ce-test-s3-b.s3.us-east-2.amazonaws.com","path":"/","accessKey":"access key","secretKey":"secret key","headers":{"x-amz-storage-class":"REDUCED_REDUNDANCY","x-amz-server-side-encryption":"AES256"}}}') # nolint
+      response <- httptest::fake_response(
+        req$url,
+        req$method,
+        content = '{
+          "s3Status":200,
+          "s3StatusMessage":"OK",
+          "s3Headers":{
+            "x-amz-id-2":"MigW9a2LcKC/DJucA0rVt2vqwwUJQJuJxlzMsmzdjTafoA9T+sMSK/Vt5paFlIPnEheq2sWXE3ekkUtopKoP8g==",
+            "x-amz-request-id":"RJT8S9PGDPJ9N0Y3",
+            "date":"Thu, 18 Aug 2022 03:41:55 GMT",
+            "etag":"\\"844da3ef3926fdfe1d5e2fc35a34d6b4\\"",
+            "server":"AmazonS3",
+            "content-length":"0"
+          },
+          "s3Response":""
+        }',
+        status_code = 200,
+        headers = list(`Content-Type` = "application/json")
+      )
+      return(response)
+    },
+    {
+      setServer("https://my.server.com")
+      resp <- getEvents(
+        "591a3b441d725115208a6fda",
+        "591a3b441d725115208a6fdc",
+        destination=list(
+          host="ce-test-s3-b.s3.us-east-2.amazonaws.com",
+          path="/",
+          accessKey="access key",
+          secretKey="secret key",
+          headers=list(
+            `x-amz-storage-class`="REDUCED_REDUNDANCY",
+            `x-amz-server-side-encryption`="AES256"
+          )
+        )
+      )
     }
   )
 })